requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.129.2
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
limits==5.8.0
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
redis==7.2.1
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
